+++ server/src/controllers/CategoryController.ts
@@ -1,0 +1,7 @@
+import { Request, Response } from 'express';
+
+const CategoryController = {
+  getCategories: async (req: Request, res: Response) => {},
+};
+
+export default CategoryController;
