+++ server/src/routes/AuthRouter.ts
@@ -1,5 +1,6 @@
 import { Router } from 'express';
-import AuthController from '../controllers/auth';
+
+import AuthController from '../controllers/AuthController';
 
 const AuthRouter = Router();
